move Prefs class to its own file
